AI/server/requirements.txt
numpy
pillow
tensorflow==2.12.0
joblib
